pyproject.toml
[tool.poetry]
name = "kegeproject"
version = "0.1.0"
description = ""
authors = ["vplotnikov16 <vplotnikov16@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.group.main.dependencies]
python = ">=3.10,<4.0"
flask = "^3.1.2"
wtforms = "^3.2.1"
flask-wtf = "^1.2.2"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.1.0"
flask-login = "^0.6.3"
flask-admin = "^2.0.2"
python-dotenv = "^1.2.1"
poetry-core = "^2.2.1"
beautifulsoup4 = "^4.10.0"

[tool.poetry.group.dev.dependencies]
pylint = ">=3.0.0"
pytest = "^7.4"
pytest-cov = "^4.0"
black = "^24.0"
isort = "^6.0"
faker = "^38.2.0"
git-filter-repo = "^2.47.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
